Moving the implementation of streaming calls from StreamingChatClientOperations to ChatClientLlmOperations to enable mocking of streaming calls at the same API level as non‑streaming ones in EmbabelMockitoIntegrationTest. Also adding support for mocking streaming calls in test cases that are based on EmbabelMockitoIntegrationTest.
Closes: #1565
